soal test tahapdua/requirements.txt
flask
Flask-Migrate
PyMySQL
flask-restful
Flask-SQLAlchemy
flask-jwt-extended